requirements.txt
Flask>=2.3.0
Flask-Mail>=0.9.1
python-dotenv>=1.0.0
pymongo>=4.3.0
requests>=2.31.0
beautifulsoup4>=4.12.2
gunicorn>=21.2.0
loguru>=0.7.0
pandas>=2.0.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
joblib>=1.3.0
